# sna/blt: An actual fast path.