--- a/src/sna/blt.c
+++ b/src/sna/blt.c
@@ -33,7 +33,6 @@
 #include <pixman.h>
 
 #include "blt_sse2.h"
-#include "blt_sse4_2.h"
 
 #if __x86_64__
 #define have_sse2() 1
@@ -586,15 +585,6 @@ void choose_memcpy_tiled_x(struct kgem *kgem, int swizzling, unsigned cpu)
 		break;
 	case I915_BIT_6_SWIZZLE_NONE:
 		DBG(("%s: no swizzling\n", __FUNCTION__));
-#if defined(sse4_2)
-		if (cpu & SSE4_2)
-		{
-			kgem->memcpy_to_tiled_x = memcpy_to_tiled_x__swizzle_0__sse42;
-			kgem->memcpy_from_tiled_x = memcpy_from_tiled_x__swizzle_0__sse42;
-			kgem->memcpy_between_tiled_x = memcpy_between_tiled_x__swizzle_0__sse42;
-		} 
-		else
-#endif
 #if defined(sse2)
 		if (cpu & SSE2)
 		{