Draft until further notice.

Before:

```
---------------------------------------------------------------------------------------------------------------------
Benchmark                                                           Time             CPU   Iterations UserCounters...
---------------------------------------------------------------------------------------------------------------------
CopyBench<shm::memory::dragons::SSE2Intel>/8                     8.26 ns         8.25 ns     85034290 bytes_per_second=925.29Mi/s
CopyBench<shm::memory::dragons::SSE2Intel>/64                    23.0 ns         22.9 ns     30609810 bytes_per_second=2.60097Gi/s
CopyBench<shm::memory::dragons::SSE2Intel>/512                    137 ns          137 ns      5156755 bytes_per_second=3.47901Gi/s
CopyBench<shm::memory::dragons::SSE2Intel>/4096                   986 ns          985 ns       715709 bytes_per_second=3.8734Gi/s
CopyBench<shm::memory::dragons::SSE2Intel>/32768                 7901 ns         7887 ns        88972 bytes_per_second=3.86918Gi/s
CopyBench<shm::memory::dragons::SSE2Intel>/262144               63245 ns        63102 ns        11077 bytes_per_second=3.86898Gi/s
CopyBench<shm::memory::dragons::SSE2Intel>/2097152             506326 ns       504417 ns         1370 bytes_per_second=3.87204Gi/s
```

After:
```
---------------------------------------------------------------------------------------------------------------------
Benchmark                                                           Time             CPU   Iterations UserCounters...
---------------------------------------------------------------------------------------------------------------------
CopyBench<shm::memory::dragons::SSE2Intel>/8                     8.61 ns         8.59 ns     80968212 bytes_per_second=887.731Mi/s
CopyBench<shm::memory::dragons::SSE2Intel>/64                    10.6 ns         10.5 ns     66428603 bytes_per_second=5.65416Gi/s
CopyBench<shm::memory::dragons::SSE2Intel>/512                   22.2 ns         22.2 ns     28525397 bytes_per_second=21.4831Gi/s
CopyBench<shm::memory::dragons::SSE2Intel>/4096                   172 ns          172 ns      4069542 bytes_per_second=22.1749Gi/s
CopyBench<shm::memory::dragons::SSE2Intel>/32768                 1778 ns         1774 ns       394876 bytes_per_second=17.199Gi/s
CopyBench<shm::memory::dragons::SSE2Intel>/262144               22946 ns        22886 ns        30420 bytes_per_second=10.6675Gi/s
CopyBench<shm::memory::dragons::SSE2Intel>/2097152             200285 ns       199388 ns         3498 bytes_per_second=9.79558Gi/s
```

Tested on a Haswell mobile workstation with an `i7-4700MQ`, with the CPU governor set to performance and turbo boost disabled.